requirements.txt
attrs==17.4.0
future==0.16.0
pluggy==0.6.0
py==1.5.2
pyglet==1.3.0
pytest==3.3.2
PyYAML==3.12
six==1.11.0
